Update the default maximum for water
The default maximum for water was way too low.
This resulted in a scaled value that made it consider
water to be about 50% as valuable as coal and 40% as valuable as copper. 

Because water is much more heavily used in recipes, this scaling caused it to prefer using lots of copper/coal/etc over water, even when only using resource weighting.